pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "envsync"
version = "0.1.0"
description = "EnvSync - 环境同步管理框架"
readme = "README.md"
authors = [{ name = "EnvSync", email = "dev@example.com" }]
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.7",
    "PyYAML>=6.0.1",
    "GitPython>=3.1.43",
    "paramiko>=3.4.0",
    "cryptography>=41.0.7",
    "rich>=13.7.0",
    "Jinja2>=3.1.3",
]

[project.scripts]
envsync = "envsync.cli:main"

[tool.setuptools.packages.find]
where = ["."]
